+++ test/sanity/sanity_windows.go
@@ -9,7 +9,7 @@ package sanity
 import (
 	"testing"
 
-	"github.com/aws/amazon-cloudwatch-agent-test/internal/common"
+	"github.com/aws/amazon-cloudwatch-agent-test/util/common"
 )
 
 func SanityCheck(t *testing.T) {
